rename to_str -> prettyprint
