+++ qiskit_optimization/problems/quadratic_expression.py
@@ -286,12 +286,12 @@ class QuadraticExpression(QuadraticProgramElement):
 
     def __repr__(self):
         # pylint: disable=cyclic-import
-        from ..translators.str import _expr2str, DEFAULT_TRUNCATE
+        from ..translators.prettyprint import _expr2str, DEFAULT_TRUNCATE
 
         return f"<LinearExpression: {_expr2str(quadratic=self, truncate=DEFAULT_TRUNCATE)}>"
 
     def __str__(self):
         # pylint: disable=cyclic-import
-        from ..translators.str import _expr2str
+        from ..translators.prettyprint import _expr2str
 
         return f"{_expr2str(quadratic=self)}"
