+++ qiskit_optimization/problems/quadratic_objective.py
@@ -203,14 +203,14 @@ class QuadraticObjective(QuadraticProgramElement):
 
     def __repr__(self):
         # pylint: disable=cyclic-import
-        from ..translators.str import _expr2str, DEFAULT_TRUNCATE
+        from ..translators.prettyprint import _expr2str, DEFAULT_TRUNCATE
 
         expr_str = _expr2str(self.constant, self.linear, self.quadratic, DEFAULT_TRUNCATE)
         return f"<QuadraticObjective: {self._sense.name.lower()} {expr_str}>"
 
     def __str__(self):
         # pylint: disable=cyclic-import
-        from ..translators.str import _expr2str
+        from ..translators.prettyprint import _expr2str
 
         expr_str = _expr2str(self.constant, self.linear, self.quadratic)
         return f"{self._sense.name.lower()} {expr_str}"
